Delete Post (Part II)
ProfilePost.jsx:
The hook usePostStore manages the deletion of posts locally.
Prevent multiple deletions if the deletion process is already ongoing.
Remove the post from the user's profile in the local store.
A success toast is shown after the post is successfully deleted.
The delete button provides visual feedback to the user when the deletion process is in progress.

userProfileStore.js:
The deletePost method removes the deleted post's ID from the user's profile.

+++ src/components/Profile/ProfilePost.jsx
@@ -19,15 +19,19 @@ const ProfilePost = ({post}) => {
   const authUser = useAuthStore(state => state.user)
   const showToast = useShowToast()
   const [isDeleting, setIsDeleting] = useState(false)
+  const deletePost = usePostStore(state => state.deletePost)
 
   const handleDeletePost = async () => {
     if (!window.confirm("Are you sure you want to delete this post?")) return 
+    if (isDeleting) return 
     try {
       const imageRef = ref(storage, `posts/${post.id}`)
       await deleteObject(imageRef)
       await deleteDoc(doc(firestore, "posts", post.id))
       const userRef = doc(firestore, "users", authUser.uid)
       await updateDoc(userRef, {posts: arrayRemove(post.id)})
+      deletePost(post.id)
+      showToast("Success", "Post deleted successfully", "success")
     } catch (error) {
       showToast("Error", error.message, "error")
     } finally {
@@ -71,7 +75,7 @@ const ProfilePost = ({post}) => {
                 </Flex>
                 {authUser?.uid === userProfile.uid && (
                   <Button _hover={{bg:"whiteAlpha.300", color:"red.600"}} borderRadius={4} p={1} size={"sm"} bg={"transparent"}
-                    onClick={handleDeletePost}>
+                    onClick={handleDeletePost} isLoading={isDeleting}>
                       <MdDelete size={20} cursor={"pointer"} />
                   </Button>
                 )}
